requirements.txt
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
openai>=1.12.0
python-docx>=1.1.0
pyahocorasick>=1.2.0
lxml>=5.0.0
schedule>=1.2.0
click>=8.1.0
python-dateutil>=2.8.0
